Add page header and footer
