+++ src/components/App/App.css
@@ -1,43 +1,70 @@
 #root {
-  --max-width: 660px;
-  margin: 0 auto;
-  text-align: center;
-
-  width: 66vw;
-  max-width: var(--max-width);
-
-  @media (max-width: 900px) {
-    width: 75vw;
-  }
-
-  @media (max-width: 768px) {
-    width: 80vw;
-  }
-
-  @media (max-width: 480px) {
-    width: 90vw;
-  }
-
-  @media (max-width: 320px) {
-    width: 98vw;
-  }
-}
-
-.App {
+  height: 100vh;
   display: flex;
   flex-direction: column;
+}
+
+.content {
+  flex: 1;
+  display: flex;
+  align-items: center;
   justify-content: center;
 }
 
-.buttonGrid {
-  width: 100%;
-  display: grid;
-  margin-top: clamp(0.125em, 2vh, 3em);
-  gap: clamp(0.125em, 2vh, 3em) clamp(0.125em, 2vw, 3em);
+.navbar {
+  background-color: var(--bg-color-2);
+  height: 50px;
+  display: flex;
+  align-items: center;
+  padding: 0 20px;
 
-  grid-template-columns: repeat(4, 1fr);
-  /* 3 column grid on small screens */
-  @media (max-width: 480px) {
-    grid-template-columns: repeat(3, 1fr);
+  & ul {
+    display: flex;
+    list-style: none;
+    margin-left: auto;
+  }
+
+  & li {
+    text-decoration: none;
+    margin: 0 10px;
+
+    &.icon {
+      & i {
+        font-size: 36;
+      }
+
+      & a {
+        cursor: pointer;
+        color: var(--text-white);
+        display: flex;
+        align-items: center;
+
+        transition: color 0.25s;
+        &:hover {
+          color: var(--purple);
+        }
+      }
+    }
+  }
+}
+
+.header-title {
+  margin: 0;
+}
+
+footer {
+  background-color: var(--bg-color-2);
+  text-align: left;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  @media (max-height: 680px) {
+    display: none;
+  }
+
+  @media (max-width: 480px) and (max-height: 740px) {
+    display: none;
   }
 }
