+++ src/components/ButtonPad/ButtonPad.css
@@ -38,8 +38,11 @@
     &:last-of-type {
       border-top-right-radius: 0px;
       border-top-left-radius: 0px;
-      padding: 0.25em;
+      padding: 0.25em 0;
       margin-top: 1px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
     }
   }
 }
