+++ src/components/Modal/HelpModal.tsx
@@ -26,7 +26,7 @@ export const HelpModal: Component<HelpModalProps> = (props) => (
   >
     <div class="grid">
       <div class="col">
-        <h1>SoundBored</h1>
+        <h2 class="pb-2">SoundBored</h2>
         <div class="text-left">
           <div class="py-2">Trigger samples by hitting the pads</div>
           <div class="py-2">Load new sounds from the sample explorer</div>
