+++ src/components/Modal/Modal.css
@@ -1,10 +1,13 @@
 .modal {
+  @media (max-width: 480px) {
+    width: 90vw;
+  }
+
   z-index: 2;
   display: flex;
   flex-direction: column;
   justify-content: center;
-  width: 50vw;
-  max-width: calc(0.75 * var(--max-width));
+  /* max-width: calc(0.75 * var(--max-width)); */
   padding: 1.3rem;
   min-height: 250px;
   position: absolute;
