+++ src/components/SampleView/SampleView.css
@@ -9,7 +9,8 @@
   position: relative;
   width: 100%;
   height: 50%;
-  background-color: var(--fg-dark);
+  /* background-color: var(--fg-dark); */
+  background-color: var(--bg-color-2);
   font-size: small;
   text-align: left;
   word-break: break-all;
