+++ src/components/index.ts
@@ -1,7 +1,8 @@
 export { App } from './App/App';
 export { ButtonPad } from './ButtonPad/ButtonPad';
-export { Modal } from './Modal/Modal';
 export { HelpModal } from './Modal/HelpModal';
+export { Modal } from './Modal/Modal';
 export { SampleControls } from './SampleControls/SampleControls';
 export { SampleEditor } from './SampleEditor/SampleEditor';
 export { SampleExplorer } from './SampleExplorer/SampleExplorer';
+export { Sampler } from './Sampler/Sampler';
