+++ src/index.css
@@ -1,41 +1,40 @@
 @import url('https://fonts.cdnfonts.com/css/avenir');
 
 :root {
+  /* Colours */
+  --text-white: rgba(255, 255, 255, 0.87);
+  --bg-color: #242424;
+  --bg-color-2: #1a1a1a;
+  --fg-dark: #1a1a1a;
+  --fg-light: #5a5a5a;
+  --purple: #646cff;
+  --orange: #da571a;
+
   font-family: Avenir, Inter, Helvetica, Arial, sans-serif;
   font-size: 16px;
   line-height: 24px;
   font-weight: 400;
 
   color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
+  color: var(--text-white);
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   -webkit-text-size-adjust: 100%;
-
-  /* Colours */
-  --bg-color: #242424;
-  --fg-dark: #1a1a1a;
-  --fg-light: #5a5a5a;
-  --purple: #646cff;
-  --orange: #da571a;
 }
 
-html,
-body {
-  width: 100%;
-  height: 100%;
-}
-
-body {
+* {
   margin: 0;
-  position: relative;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  height: 100vh;
   display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
+  flex-direction: column;
   background-color: var(--bg-color);
 }
